Consolidate playlists routes tests into one file
